% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valuefunction.R
\name{ActionValueNetwork}
\alias{ActionValueNetwork}
\title{Action Value Network}
\description{
Neural network representing the action value function Q.
}
\examples{
library(keras)
model = keras_model_sequential()
model \%>\% layer_dense(20, input_shape = 10, activation = "relu")
model \%>\% layer_dense(4, activation = "softmax")
keras::compile(model, loss = "mae", optimizer = keras::optimizer_sgd(lr = 0.4))

val = ActionValueNetwork$new(model)
val$predictQ(matrix(1, ncol = 10))
val$train(matrix(1, ncol = 10), target = matrix(c(-1, 2, 4, -3), ncol = 4))
}
