% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valuefunction.R
\name{ActionValueTable}
\alias{ActionValueTable}
\title{Action Value Table}
\description{
Table representing the action value function Q.
}
\examples{
val = makeValueFunction("table", n.states = 20L, n.actions = 4L)
}
