% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glie_monte_carlo.R
\name{controlMonteCarlo}
\alias{controlMonteCarlo}
\title{On-policy Monte Carlo Control}
\usage{
controlMonteCarlo(policy, envir, discount.factor)
}
\arguments{
\item{policy}{a policy specified as a probability matrix (states x actions)}

\item{envir}{the environment, an R6 class. See also \link{makeEnvironment}.}

\item{discount.factor}{scalar numeric, discounting future rewards}
}
\value{
optimal policy and optimal action value function
}
\description{
GLIE Monte Carlo control. Evaluate policy, then epsilon-greedy policy
improvement.
}
