% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eligibility.R
\name{Eligibility}
\alias{Eligibility}
\alias{eligibility}
\title{Eligibility traces}
\arguments{
\item{lambda}{[\code{numeric(1)} in (0, 1)] \cr Trace decay parameter.}

\item{traces}{[\code{character(1)}] \cr Type of eligibility trace update. One of \code{c("replace", "accumulate")}.}
}
\description{
Eligibility traces.
}
\details{
Algorithms supporting eligibility traces:
\itemize{
\item \link{QLearning}
}
}
\examples{
alg = makeAlgorithm("qlearning", lambda = 0.8, traces = "accumulate")
}
