% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/environment.R
\name{Environment}
\alias{Environment}
\title{Reinforcement Learning Environment}
\arguments{
\item{step}{[\code{function(self, action)}] \cr
Custom step function.}

\item{reset}{[\code{function()}] \cr
Custom reset function.}

\item{visualize}{[\code{function(self)}] \cr
Optional custom visualize function.}
}
\description{
Reinforcement Learning Environment
}
\section{Usage}{

\code{Environment$new(step, reset, visualize)}
}

\section{Methods}{

\code{$new()} Initializes a new environment.

\code{$reset()} Starts a new episode.
Returns the \code{state}.

\code{$step(action)} Take action in environment.
Returns a list with \code{state}, \code{reward}, \code{done}.

\code{$visualize()} Visualizes environment.
}

\examples{
step = function(self, action) {
  state = list(mean = action + rnorm(1), sd = runif(1))
  reward = rnorm(1, state[[1]], state[[2]])
  done = FALSE
  list(state, reward, done)
}

reset = function() {
  state = list(mean = 0, sd = 1)
  state
}

env = Environment$new(step, reset)
env$reset()
env$step(100)
}
