% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frozenlake.R
\docType{class}
\name{FrozenLake}
\alias{FrozenLake}
\title{Frozen Lake Environment}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
FrozenLake
}
\description{
The agent controls the movement of a character in a grid world. Some tiles of
the grid are walkable, and others lead to the agent falling into the water.
Additionally, the movement direction of the agent is uncertain and only
partially depends on the chosen direction. The agent is rewarded for finding
a walkable path to a goal tile. See \link{https://gym.openai.com/envs/FrozenLake-v0}
for details.
}
\details{
To use this you must have installed OpenAI Gym und gym-http-api. To start a
session run python gym_http_server.py in your command line.
}
\section{Methods}{
 \describe{ \item{\code{FrozenLake$new()}}{Creates a new
\code{FrozenLake} environment.} \item{\code{FrozenLake$step(state,
  action, render = TRUE)}}{ Takes a step in the environment given a state and
an action, returns the next state and reward. If render = TRUE the
environment will be rendered.}
\item{\code{FrozenLake$setEpisodeOverFalse()}}{Resets the
\code{episode.over} flag of the FrozenLake environment. Useful when
starting a new episode.} }
}

\keyword{datasets}
