% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/policy.R
\name{GaussianPolicy}
\alias{GaussianPolicy}
\title{Gaussian Policy}
\description{
Gaussian Policy
}
\section{Usage}{

\code{GaussianPolicy$new()}
}

\section{Methods}{

\code{$sampleAction(policy)} Sample action from policy probabilities.
}

\examples{
pol = GaussianPolicy$new()
pol$sampleAction(mean = 10, sd = 1)
}
