% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessor_functions.R
\name{getEligibilityTraces}
\alias{getEligibilityTraces}
\title{Get eligibility traces}
\usage{
getEligibilityTraces(agent)
}
\arguments{
\item{agent}{[Agent] \cr An agent created by \link{makeAgent}.}
}
\value{
A matrix with the eligibility traces.
}
\description{
Returns the eligibility traces of the agent.
}
