% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessor_functions.R
\name{getReplayMemory}
\alias{getReplayMemory}
\title{Get replay memory.}
\usage{
getReplayMemory(agent)
}
\arguments{
\item{agent}{[Agent] \cr An agent created by \link{makeAgent}.}
}
\value{
A list containing the experienced observations, actions and rewards.
}
\description{
Returns the replay memory of the agent.
}
