% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessor_functions.R
\name{getStateValues}
\alias{getStateValues}
\title{Get state values.}
\usage{
getStateValues(action.vals)
}
\arguments{
\item{action.vals}{[\code{matrix}] \cr Action value matrix.}
}
\description{
Get state value function from  action value function.
}
