% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessor_functions.R
\name{getValueFunction}
\alias{getValueFunction}
\title{Get weights of value function.}
\usage{
getValueFunction(agent)
}
\arguments{
\item{agent}{[Agent] \cr An agent created by \link{makeAgent}.}
}
\value{
For a value function table this will return a matrix, for a neural
network a list with the weights of the layers.
}
\description{
Returns the weights of the value function representation of the agent.
}
