% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gridworld_step.R
\name{gridworld_step}
\alias{gridworld_step}
\title{Take a step in a gridworld}
\usage{
gridworld_step(state, action, gridworld)
}
\arguments{
\item{state}{scalar integer: the current state}

\item{action}{character: the action, one of ("left", "right", "up", "down")}

\item{gridworld}{object returned by \code{\link{gridworld}}}
}
\value{
list with next state, reward and a flag if episode is finished, i.e. the new state is
a terminal state
}
\description{
Given a state and an action in a gridworld, the next state and reward are returned.
}
\details{
Possible actions include going left, right, down or up. If an action would take you off
the grid, you remain in the previous state. For each step you get a reward of -1, until you reach
into a terminal state.
}
\examples{
set.seed(27)
grid = gridworld()
actions = c("left", "right", "up", "down")
# specify the initial state
states = 3
rewards = numeric(0)
sampled.actions = character(0)
episode.over = FALSE
i = 1

while(episode.over == FALSE) {
  sampled.actions = append(sampled.actions, sample(actions, size = 1))
  step = gridworld_step(states[i], sampled.actions[i], grid)
  rewards = append(rewards, step[["reward"]])
  states = rbind(states, step[["state"]])
  episode.over = step[["episode.over"]]
  i = i + 1
}

print(rewards)
print(states)

}
\references{
Gridworld example from Sutton & Barto, chapter 4
}
\seealso{
\code{\link{gridworld}}
}
