% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/environment_gym.R
\name{GymEnvironment}
\alias{GymEnvironment}
\title{Gym Environment}
\arguments{
\item{gym}{[\code{character(1)}] \cr
Name of gym environment, e.g. \code{"CartPole-v0"}.}
}
\description{
Reinforcement learning environment from OpenAI Gym.
}
\details{
For OpenAI gym environments have a look at \url{https://gym.openai.com/envs}.
}
\section{Installation}{

For installation of the python package \code{gym} see
\url{https://github.com/openai/gym#installation}.
Then install the R package \code{reticulate}.
}

\section{Usage}{

\code{makeEnvironment("Gym", gym.name)}
}

\section{Methods}{

 \describe{
\item{\code{$close()}}{
Close visualization window.
 }
 }


\describe{
\item{\code{$step(action)}}{
Take action in environment.
Returns a list with \code{state}, \code{reward}, \code{done}.
 }
\item{\code{$reset()}}{
  Resets the \code{done} flag of the environment and returns an initial state.
  Useful when starting a new episode.
}
\item{\code{$visualize()}}{
  Visualizes the environment (if there is a visualization function).
}
}
}

\examples{
\dontrun{
# Create an OpenAI Gym environment.
# Make sure you have Python, gym and reticulate installed.
env = makeEnvironment("Gym", "MountainCar-v0")
env$reset()
env$close()
}
}
