% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/actorCritic.R
\name{identity2}
\alias{identity2}
\title{Identity Function}
\usage{
identity2(x, ...)
}
\arguments{
\item{x}{[\code{R object}] \cr
This argument will be returned.}

\item{...}{Other arguments, which are not used.}
}
\description{
A trivial identity function returning its argument.
In comparison to \code{\link{identity}} it accepts more than one argument,
but always returns the first argument.
}
\examples{
identity2(10, "a")
}
