% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sarsa_approx.R
\name{make_feature_vector}
\alias{make_feature_vector}
\title{Make feature vector using tiles}
\usage{
make_feature_vector(state, state.space.bounds, n.grid)
}
\arguments{
\item{state}{length-two list: the state observation}

\item{state.space.bounds}{length-two list: the state space bounds}

\item{n.grid}{integer scalar: sqrt of number of grid cells}
}
\value{
one-hot feature vector
}
\description{
Tiles: Grid across the 2-dimensional state space: each observation
will be encoded as a one-hot vector.
}
