% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qnetwork.R
\name{make_one_hot_vector}
\alias{make_one_hot_vector}
\title{Make one-hot vector}
\usage{
make_one_hot_vector(hot, len, matrix = TRUE)
}
\arguments{
\item{hot}{scalar integer: which entry is 1}

\item{len}{scalar integer: the length of the one-hot vector}

\item{matrix}{logical: should the output be a matrix with one row}
}
\value{
one_hot vector or matrix
}
\description{
Make one-hot vector
}
