% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agent.R
\name{makeAgent}
\alias{makeAgent}
\title{Create Agent.}
\usage{
makeAgent(policy, val.fun = NULL, algorithm = NULL, preprocess = identity,
  experience.replay = NULL, eligibility.traces = NULL)
}
\arguments{
\item{policy}{[\code{Policy}] \cr A policy created by \link{makePolicy}.}

\item{val.fun}{[\code{ActionValueTable}] \cr Value function representation.}

\item{algorithm}{[\code{Algorithm}] \cr An algorithm.}
}
\description{
Create Agent.
}
