% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithm.R
\name{makeAlgorithm}
\alias{makeAlgorithm}
\title{Make reinforcement learning algorithm.}
\usage{
makeAlgorithm(class, ...)
}
\arguments{
\item{class}{[\code{character(1)}] \cr Algorithm.}

\item{...}{[\code{any}] \cr Arguments passed on to algorithm.}
}
\description{
Make reinforcement learning algorithm.
}
