% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tiles.R
\name{makeNHot}
\alias{makeNHot}
\title{Make n hot vector}
\usage{
makeNHot(x, len, out = "matrix")
}
\arguments{
\item{x}{[\code{integer}] \cr Which features are active?}

\item{len}{[\code{integer(1)}] \cr Length of the feature vector.}

\item{out}{[\code{character(1)}] \cr Format of the output. Can be a vector or a matrix.}
}
\value{
[\code{matrix(1, len)}] A one-row matrix with \code{len} columns with every entry 0 
  except the columns specified by \code{x}.
}
\description{
Make n hot vector
}
\examples{
makeNHot(c(1, 3), 5)
makeNHot(c(1, 3), 5, out = "vector")
}
