% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/policy.R
\name{makePolicy}
\alias{makePolicy}
\alias{Policy}
\title{Create policy.}
\usage{
makePolicy(class = "random", ...)
}
\arguments{
\item{class}{[\code{character(1)}] \cr
Class of policy. See Section Policies for available policies.}

\item{...}{[\code{any}] \cr
Arguments passed on to the policy class.}
}
\description{
Reinforcement learning policies.
}
\section{Policies}{

\itemize{
\item \link{RandomPolicy}
\item \link{GreedyPolicy}
\item \link{EpsilonGreedyPolicy}
\item \link{SoftmaxPolicy}
}
}

\examples{
policy = makePolicy("random")
policy = makePolicy("epsilon.greedy", epsilon = 0.1)
}
