% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experience_replay.R
\name{makeReplayMemory}
\alias{makeReplayMemory}
\alias{experiencereplay}
\title{Experience Replay}
\usage{
makeReplayMemory(size = 100L, batch.size = 16L)
}
\arguments{
\item{size}{[\code{integer(1)}] \cr Size of replay memory.}

\item{batch.size}{[\code{integer(1)}] \cr Batch size.}
}
\description{
Create replay memory for experience replay.
}
