% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valuefunction.R
\name{makeValueFunction}
\alias{makeValueFunction}
\title{Value Function Representation}
\usage{
makeValueFunction(class, ...)
}
\arguments{
\item{class}{[\code{character(1)}] \cr Class of value function approximation.}

\item{...}{[\code{any}] \cr Arguments passed to subclass.}
}
\description{
Value Function Representation
}
\section{Representations}{

\itemize{
\item \link{ValueTable}
}
}

