% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeWindyGridworld.R
\name{makeWindyGridworld}
\alias{makeWindyGridworld}
\title{Make Windy Gridworld environment}
\usage{
makeWindyGridworld(shape = c(7L, 10L), terminal.states = 37L, wind = c(0L,
  0L, 0L, 1L, 1L, 1L, 2L, 2L, 1L, 0L))
}
\arguments{
\item{shape}{length-two integer, shape of the grid}

\item{terminal.states}{integer vector of the terminal states in the
gridworld, states are numerated starting with 0 with increasing
number from left to right}

\item{wind}{integer vector specifying the strength of the upward
wind in each cell.}
}
\value{
R6 class
}
\description{
Windy Gridworld problem for reinforcement learning. Actions include
going left, right, up and down. In each column the wind pushes you up a
specific number of steps (for thenext action). If an action would
take you off the grid, you remain in the previous state. For each step you
get a reward of -1, until you reach into a terminal state.
}
\examples{
grid = makeWindyGridworld()
grid = makeEnvironment(transition.array = grid$transition.array, 
  reward.matrix = reward.matrix, initial.state = 30L)
}
\references{
Sutton and Barto (Book draft 2016): Reinforcement Learning: An Introduction
}
\seealso{
\link{makeEnvironment}
}
