% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MonteCarloES.R
\name{MonteCarloES}
\alias{MonteCarloES}
\title{Monte Carlo Control with Exploring Starts}
\usage{
MonteCarloES(policy, envir, discount.factor)
}
\arguments{
\item{policy}{a policy specified as a probability matrix (states x actions)}

\item{envir}{the environment, a function returning the next state and reward given an action}

\item{discount.factor}{scalar numeric, discounting future rewards}
}
\value{
the optimal policy and optimal action value function q
}
\description{
Monte Carlo Control with Exploring Starts
}
