% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/environment_mountaincar.R
\docType{data}
\name{MountainCar}
\alias{MountainCar}
\alias{MountainCar}
\alias{MountainCarContinuous}
\title{Mountain Car}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\description{
The classical mountain car problem for reinforcement learning.
}
\details{
The classical Mountain Car task the action is one of \{0, 1, 2\},
in the continuous version the action is in [-1, 1].
}
\section{Usage}{

\code{MountainCar$new()} \cr
\code{MountainCarContinuous$new()}
}

\section{Methods}{

\code{$new()} Initializes a new environment.

\code{$reset()} Starts a new episode.
Returns the \code{state}.

\code{$step(action)} Take action in environment.
Returns a list with \code{state}, \code{reward}, \code{done}.

\code{$visualize()} Visualizes environment.
}

\examples{
env = MountainCar$new()
env$reset()
env$step(1L)

env = MountainCarContinuous$new()
env$reset()
env$step(0.62)
}
\keyword{datasets}
