% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/policy.R
\name{Policy}
\alias{Policy}
\title{Policy}
\description{
Reinforcement learning policies.
}
\section{Policies}{

\itemize{
\item \link{RandomPolicy}
\item \link{GreedyPolicy}
\item \link{EpsilonGreedyPolicy}
\item \link{GaussianPolicy}
\item \link{SoftmaxPolicy}
}
}

\section{Methods}{

\code{$sampleAction(policy)} Sample action from policy probabilities.
}

