% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qlearning.R
\name{qlearning}
\alias{qlearning}
\title{Q-Learning}
\usage{
qlearning(behaviour.policy, envir)
}
\arguments{
\item{behaviour.policy}{matrix: the behaviour policy}

\item{envir}{the environment, a function returning the next state and reward given an action}
}
\value{
the optimal policy and optimal action value function
}
\description{
Off-policy control.
}
