% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reinforcelearn.R
\docType{package}
\name{reinforcelearn}
\alias{reinforcelearn}
\alias{reinforcelearn-package}
\title{Reinforcement Learning.}
\description{
Implementations of reinforcement learning algorithms and environments.
}
\section{Environments}{

\itemize{
\item \link{Environment}
\item \link{MdpEnvironment}
\item \link{Gridworld}
\item \link{WindyGridworld}
\item \link{CliffWalking}
\item \link{MountainCar}
\item \link{GymEnvironment}
}
}

