% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sarsa.R
\name{sarsa}
\alias{sarsa}
\title{SARSA(\lambda)}
\usage{
sarsa(policy, envir, lambda, discount.factor)
}
\arguments{
\item{policy}{a policy specified as a probability matrix (states x actions)}

\item{envir}{the environment, a function returning the next state and reward given an action}

\item{lambda}{scalar integer between 0 and 1}

\item{discount.factor}{scalar numeric, discounting future rewards}
}
\value{
optimal policy and optimal action value function
}
\description{
SARSA(\lambda)
}
