% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/policy.R
\name{SoftmaxPolicy}
\alias{SoftmaxPolicy}
\title{Softmax Policy}
\description{
Softmax Policy
}
\section{Usage}{

\code{SoftmaxPolicy$new()}
}

\section{Methods}{

\code{$sampleAction(policy)} Sample action from policy probabilities.
}

\examples{
pol = SoftmaxPolicy$new()
(probs = pol$getActionProbs(matrix(c(1:3), ncol = 3)))
pol$sampleAction(probs)
}
