% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bandit.R
\name{solveBandit}
\alias{solveBandit}
\title{Solve multi-armed bandit}
\usage{
solveBandit(bandit, n.episodes = 10L, action.selection = c("epsilon-greedy",
  "greedy", "UCB"), epsilon = 0.1, epsilon.decay = NULL,
  initial.value = 0, initial.visits = 0L, c = 2, ...)
}
\arguments{
\item{bandit}{R6 class bandit}

\item{n.episodes}{scalar integer: number of episodes}

\item{action.selection}{scalar character: which method to use for
action selection, e.g. "epsilon-greedy", "greedy" or "UCB"}

\item{epsilon}{scalar numeric: ratio of random exploration in
epsilon-greedy action selection}

\item{epsilon.decay}{scalar numeric between 0 and 1: decay epsilon
every 100 episodes by this factor}

\item{initial.value}{scalar numeric: initial values for the action
values Q, set this to the maximal
possible reward to encourage exploration (optimistic initialization)}

\item{initial.visits}{scalar integer: set this to a high number to
encourage exploration (together with a high initial.value)}

\item{...}{arguments passed on to action selection algorithm}
}
\value{
numeric vector: the action values for the arms of the bandit
}
\description{
Solve multi-armed bandit
}
\examples{
ExampleBandit = bandit$new()
solveBandit(ExampleBandit, n.episodes = 1000, 
  action.selection = "greedy")
solveBandit(ExampleBandit, n.episodes = 1000, 
  action.selection = "epsilon-greedy", epsilon = 0.5)
solveBandit(ExampleBandit, n.episodes = 1000, 
  action.selection = "greedy", 
  initial.value = 5, initial.visits = 100)
solveBandit(ExampleBandit, n.episodes = 1000, 
  action.selection = "UCB", c = 2)
# true values: 1, 2, 2.5, 4
}
