% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valuefunction.R
\name{ValueFunction}
\alias{ValueFunction}
\title{Value Function Representation}
\description{
Value Function Representation
}
\section{Representations}{

\itemize{
\item \link{ActionValueNetwork}
\item \link{ActionValueTable}
}
}

