% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/windyGridworld.R
\name{windyGridworld}
\alias{windyGridworld}
\title{Windy Gridworld}
\format{[\code{list(2)}] \cr
  List with the transition array and reward matrix.}
\description{
Windy Gridworld problem for reinforcement learning. Actions include
going left, right, up and down. In each column the wind pushes you up a
specific number of steps (for the next action). If an action would
take you off the grid, you remain in the previous state. For each step you
get a reward of -1, until you reach into a terminal state.
}
