% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valuefunction.R
\name{ActionValueTable}
\alias{ActionValueTable}
\title{Action Value Table}
\description{
Table representing the action value function Q.
}
\examples{
val = ActionValueTable$new(n.states = 3, n.actions = 4)
val$predictQ(matrix(1, ncol = 3))
val$train(2, target = matrix(c(-1, 2, 4, -3), ncol = 4))
val$Q
}
