% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bandit.R
\docType{data}
\name{bandit}
\alias{bandit}
\title{Example Bandit}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
bandit
}
\value{
R6 class
}
\description{
Example Bandit
}
\keyword{datasets}
