% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/policy.R
\name{EpsilonGreedyPolicy}
\alias{EpsilonGreedyPolicy}
\alias{GreedyPolicy}
\title{Epsilon Greedy Policy}
\arguments{
\item{epsilon}{[\code{numeric(1) in [0, 1]}] \cr
Ratio of random exploration in epsilon-greedy action selection.}
}
\description{
Epsilon Greedy Policy
}
\section{Usage}{

\code{makePolicy("epsilon.greedy", epsilon)} \cr
\code{makePolicy("greedy")}
}

\examples{
policy = makePolicy("epsilon.greedy", epsilon = 0.1)
(probs = policy$getActionProbs(matrix(c(1:3), ncol = 3), n.actions = 3))
policy$sampleAction(probs)
}
