% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluatePolicy.R
\name{evaluatePolicy}
\alias{evaluatePolicy}
\title{Policy Evaluation}
\usage{
evaluatePolicy(policy, envir, discount.factor = 1, epsilon = 1e-04)
}
\arguments{
\item{policy}{a policy specified as a probability matrix (states x actions)}

\item{envir}{the environment, a function returning the next state and reward given an action}

\item{discount.factor}{scalar numeric, discounting future rewards}

\item{epsilon}{scalar numeric, algorithm stops when improvement is smaller than epsilon}
}
\value{
the state value function v
}
\description{
Evaluate a given policy in an environment.
}
\details{
The algorithm runs until the improvement in the value function in two subsequent steps
is smaller than epsilon.
}
\examples{
# Define uniform random policy, take each action with probability 0.25
grid = gridworld$new()
random.policy = matrix(1 / grid$n.actions, nrow = grid$n.states, 
  ncol = grid$n.actions)

# Evaluate given policy for gridworld example
v = evaluatePolicy(random.policy, grid)

# Compare results with expected result
v.expected = c(0, -14, -20, -22, -14, -18, -20, -20,
               -20, -20, -18, -14, -22, -20, -14, 0)
all.equal(v, v.expected, tolerance = 0.1)
}
