% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gridworld.R
\docType{class}
\name{gridworld}
\alias{gridworld}
\title{Gridworld environment}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
#grid = gridworld$new()
}
\description{
Simple gridworld for reinforcement learning. With the step method given a
state and an action in a gridworld, the next state and reward are returned.
}
\details{
The states are enumerated as follows (example 4x4 grid):
\tabular{rrrr}{
1 \tab 2 \tab 3 \tab 4 \cr
5 \tab 6 \tab 7 \tab 8 \cr
9 \tab 10 \tab 11 \tab 12 \cr
13 \tab 14 \tab 15 \tab 16 \cr
}
So a board position could look like this (T: terminal state, x: current state):
\tabular{rrrr}{
T \tab o \tab o \tab o \cr
o \tab o \tab o \tab o \cr
o \tab x \tab o \tab o \cr
o \tab o \tab o \tab T \cr
}
Possible actions include going left, right, down or up. If an action would
take you off the grid, you remain in the previous state. For each step you
get a reward of -1, until you reach into a terminal state.
}
\section{Methods}{
 \describe{ \item{\code{gridworld$new(shape, 
  terminal.states, initial.state)}}{Creates a new \code{gridworld} with a
specific \code{shape}, which is a length-two integer, e.g. \code{c(4, 4)}.
\code{terminal.states} is an integer vector of the terminal states in the
gridworld. Default is \code{c(1, 16)}. initial.state sppecifies the
initial.states, an integer vector.} }
}

\examples{
grid = gridworld$new(shape = c(4, 4), terminal.states = c(1, 16), 
  initial.state = 1:14)
}
\references{
Gridworld example from Sutton & Barto, chapter 4
}
\seealso{
\link{makeEnvironment}
}
\keyword{datasets}
