% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gridworld.R
\name{gridworld}
\alias{gridworld}
\title{Gridworld environment}
\usage{
gridworld(shape = c(4, 4), terminal.states = c(1, 16))
}
\arguments{
\item{shape}{length-two integer vector: the shape of the grid, e.g. (4, 4)}

\item{terminal.states}{integer vector of terminal states}
}
\value{
list with a 3-dimensional array containing the probability transition matrices
 for each action,  a reward matrix (states x actions) and the possibel actions as a vector.
}
\description{
Simple gridworld for reinforcement learning.
}
\details{
Possible actions include going left, right, down or up. If an action would take you off
the grid, you remain in the previous state. For each step you get a reward of -1, until you reach
into a terminal state.

The states are enumerated as follows (example 4x4 grid):
\tabular{rrrr}{
 1 \tab 2 \tab 3 \tab 4 \cr
 5 \tab 6 \tab 7 \tab 8 \cr
 9 \tab 10 \tab 11 \tab 12 \cr
 13 \tab 14 \tab 15 \tab 16 \cr
}
So a board position could look like this (T: terminal state, x: current state):
\tabular{rrrr}{
 T \tab o \tab o \tab o \cr
 o \tab o \tab o \tab o \cr
 o \tab x \tab o \tab o \cr
 o \tab o \tab o \tab T \cr
}
}
\examples{
# Create 3x5 gridworld with one terminal state in the top right corner of the grid.
gridworld(shape = c(3, 5), terminal.states = 5)
}
\references{
Gridworld example from Sutton & Barto, chapter 4
}
\seealso{
\code{\link{gridworld_step}}
}
