% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gridworld.R
\name{gridworld}
\alias{gridworld}
\title{Gridworld}
\format{[\code{list(2)}] \cr 
  List with the transition array and reward matrix.}
\description{
Simple gridworld environment for reinforcement learning.
Grid of shape 4x4 with two terminal states in the upper left and
bottom right of the grid. Episodes start in any of the nonterminal
states. The possible actions include going left, right, up and down.
}
