% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/environment_gym.R
\name{GymEnvironment}
\alias{GymEnvironment}
\title{Gym Environment}
\arguments{
\item{gym}{[\code{character(1)}] \cr
Name of gym environment, e.g. \code{"CartPole-v0"}.}
}
\description{
Reinforcement learning environment from OpenAI Gym.
}
\details{
For OpenAI gym environments have a look at \url{https://gym.openai.com/envs}.
}
\section{Installation}{

For installation of the python package \code{gym} see
\url{https://github.com/openai/gym#installation}.
Then install the R package \code{reticulate}.
}

\section{Usage}{

\code{GymEnvironment$new(gym.name)}
}

\section{Methods}{

\code{$close()} Close visualization window.


\code{$new()} Initializes a new environment.

\code{$reset()} Starts a new episode.
Returns the \code{state}.

\code{$step(action)} Take action in environment.
Returns a list with \code{state}, \code{reward}, \code{done}.

\code{$visualize()} Visualizes environment.
}

\examples{
\dontrun{
env = GymEnvironment$new("MountainCar-v0")
env$reset()
env$close()
}
}
