% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iteratePolicy.R
\name{iteratePolicy}
\alias{iteratePolicy}
\title{Policy Iteration}
\usage{
iteratePolicy(policy, envir, discount.factor)
}
\arguments{
\item{policy}{a policy specified as a probability matrix (states x actions)}

\item{envir}{the environment, a function returning the next state and reward given an action}

\item{discount.factor}{scalar numeric, discounting future rewards}
}
\value{
the optimal policy and optimal state value function
}
\description{
Find optimal policy by dynamic programming. Iterate between evaluating a
given policy (until convergence) and improving the policy by a greedy update.
Converges to the optimal policy.
}
