% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iterateValue.R
\name{iterateValue}
\alias{iterateValue}
\title{Value Iteration}
\usage{
iterateValue(policy, envir, discount.factor)
}
\arguments{
\item{policy}{numeric matrix: a policy specified as a probability matrix (states x actions)}

\item{envir}{an R6 class: the reinforcement learning environment created by \link{makeEnvironment}.}

\item{discount.factor}{scalar numeric, discounting future rewards}
}
\value{
the optimal policy and optimal state value function
}
\description{
Find optimal policy by dynamic programming. Iterate between evaluating a
given policy (only one step), then improving the policy by a greedy update.
Converges to the optimal policy.
}
