% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithm.R
\name{makeAlgorithm}
\alias{makeAlgorithm}
\title{Make reinforcement learning algorithm.}
\usage{
makeAlgorithm(class, ...)
}
\arguments{
\item{class}{[\code{character(1)}] \cr Algorithm. One of \code{c("qlearning")}.}

\item{...}{[\code{any}] \cr Arguments passed on to the specific algorithm.}
}
\description{
Make reinforcement learning algorithm.
}
\section{Representations}{

\itemize{
\item \link{QLearning}
}
}

\examples{
alg = makeAlgorithm("qlearning")
}
