% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithm.R
\name{makeAlgorithm}
\alias{makeAlgorithm}
\title{Make reinforcement learning algorithm.}
\usage{
makeAlgorithm(class, args = list(), ...)
}
\arguments{
\item{class}{[\code{character(1)}] \cr Algorithm. One of \code{c("qlearning")}.}

\item{args}{[\code{list}] \cr Optional list of named arguments passed on to the
subclass. The arguments in ... take precedence over values in this list.
We strongly encourage you to use one or the other to pass arguments
to the function but not both.}

\item{...}{[\code{any}] \cr Optional named arguments passed on to the subclass. Alternatively
these can be given using the \code{args} argument.}
}
\description{
Make reinforcement learning algorithm.
}
\section{Representations}{

\itemize{
\item \link{QLearning}
}
}

\examples{
alg = makeAlgorithm("qlearning")
}
