% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valuefunction.R
\name{makeValueFunction}
\alias{makeValueFunction}
\title{Value Function Representation}
\usage{
makeValueFunction(class, ...)
}
\description{
Value Function Representation
}
\section{Representations}{

\itemize{
\item \link{ActionValueTable}
}
}

