% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WindyGridworld.R
\name{makeWindyGridworld}
\alias{makeWindyGridworld}
\title{Make Windy Gridworld environment}
\usage{
makeWindyGridworld(shape = c(7L, 10L), terminal.states = 37L, wind = c(0L,
  0L, 0L, 1L, 1L, 1L, 2L, 2L, 1L, 0L))
}
\arguments{
\item{shape}{[\code{integer(2)}] \cr 
Shape of the gridworld}

\item{terminal.states}{[\code{integer}] \cr 
Terminal states in the 
gridworld, states are numerated starting with 0 with increasing 
number from left to right}

\item{wind}{[\code{integer}] \cr 
Strength of the upward wind in each cell.}
}
\value{
[\code{R6 class}] \cr
  Returns the windy gridworld environment.
}
\description{
Windy Gridworld problem for reinforcement learning. Actions include
going left, right, up and down. In each column the wind pushes you up a
specific number of steps (for the next action). If an action would
take you off the grid, you remain in the previous state. For each step you
get a reward of -1, until you reach into a terminal state.
}
\examples{
grid = makeWindyGridworld()
grid = makeEnvironment(transition.array = grid$transition.array, 
  reward.matrix = grid$reward.matrix, initial.state = 30L)
}
\references{
Sutton and Barto (Book draft 2017): Reinforcement Learning: An Introduction
}
\seealso{
\code{\link{makeEnvironment}}
}
