% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MonteCarloES.R
\name{MonteCarloES}
\alias{MonteCarloES}
\title{Monte Carlo Control with Exploring Starts}
\usage{
MonteCarloES(envir, n.episodes = 100, discount.factor = 1)
}
\arguments{
\item{envir}{the environment, an R6 class. See also \code{\link[=envir]{envir()}}.}

\item{n.episodes}{scalar integer: the number of episodes}

\item{discount.factor}{scalar numeric, discounting future rewards}
}
\value{
the optimal policy and optimal action value function q
}
\description{
Monte Carlo Control with Exploring Starts
}
\examples{
set.seed(123)
grid = gridworld$new()
  
# Estimate optimal action value function with Monte Carlo Exploring Starts
# Q = MonteCarloES(grid)
}
