% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mountaincar.R
\name{MountainCar}
\alias{MountainCar}
\title{Mountain Car}
\usage{
MountainCar()
}
\description{
Mountain Car
}
\examples{
m = MountainCar()
m$reset()
m$step(1)

}
\references{
Sutton and Barto (2017): Reinforcement Learning: An Introduction
}
