% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mountaincar.R
\name{mountainCar}
\alias{mountainCar}
\title{Mountain Car}
\usage{
mountainCar(action.space = "Discrete")
}
\arguments{
\item{action.space}{[\code{character(1)}] \cr
Can be \code{"Discrete"} or \code{"Continuous"}. 
The classical Mountain Car task has three discrete actions, 
the continuous version one action in a box from -1 to 1.}
}
\description{
Mountain Car
}
\examples{
# Classic Mountain Car task
m = mountainCar()
m$reset()
m$step(1)
print(m)

# Mountain Car with continuous action space
m = mountainCar(action.space = "Continuous")
m$reset()
m$step(0.27541)
print(m)

}
\references{
Sutton and Barto (2017): Reinforcement Learning: An Introduction
}
