% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mountainCar.R
\docType{class}
\name{mountainCar}
\alias{mountainCar}
\title{Mountain Car environment}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
mountainCar
}
\description{
Classic mountain car control problem from the reinforcement learning literature.
Goal is to drive a car up a steep hill. It can only be achieved by rolling
backwards up the hill, builing up momentum and then going forwards.
The state space is two-dimensional (position and velocity), the action space
is discrete (throttle forward +1, throttle backwards -1, zero throttle 0).
}
\details{
To use this you must have installed OpenAI Gym und gym-http-api.
To start a session run python gym_http_server.py in your command line.
}
\section{Methods}{
 \describe{
\item{\code{mountainCar$new()}}{Creates a new \code{mountainCar} environment.}
\item{\code{mountainCar$step(state, action, render = TRUE)}}{
Takes a step in the environment given a state and an action,
returns the next state and reward. If render = TRUE the environment will be rendered.}
\item{\code{mountainCar$setEpisodeOverFalse()}}{Resets the
\code{episode.over} flag of the mountainCar environment.
Useful when starting a new episode.}
}
}

\references{
Mountain Car example from Sutton & Barto, chapter 10
}
\keyword{datasets}
