% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithm.R
\name{QLearning}
\alias{QLearning}
\alias{qlearning}
\title{Q-Learning}
\arguments{
\item{lambda}{[\code{numeric(1)} in (0, 1)] \cr Trace decay parameter.}

\item{traces}{[\code{character(1)}] \cr Type of eligibility trace update. One of \code{c("replace", "accumulate")}.}
}
\description{
Q-Learning algorithm.
}
\details{
To use eligibility traces specify \code{lambda} and \code{traces}.
}
\section{Usage}{

\code{makeAlgorithm("qlearning", lambda, traces)}
}

\examples{
alg = makeAlgorithm("qlearning", lambda = 0.8, traces = "accumulate")
}
\seealso{
\link{Eligibility}
}
