% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/policy.R
\name{RandomPolicy}
\alias{RandomPolicy}
\title{Random Policy}
\description{
Random Policy
}
\section{Usage}{

\code{RandomPolicy$new()}
}

\section{Methods}{

\code{$sampleAction(policy)} Sample action from policy probabilities.
}

\examples{
pol = RandomPolicy$new()
(probs = pol$getActionProbs(n.actions = 4))
(probs = pol$getActionProbs(n.actions = 5))
pol$sampleAction(probs)
}
