% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reinforcelearn.R
\docType{package}
\name{reinforcelearn}
\alias{reinforcelearn}
\alias{reinforcelearn-package}
\title{Reinforcement Learning in R.}
\description{
Implementations of reinforcement learning algorithms.
}
\section{Environments}{

\itemize{
\item \link{makeEnvironment}
\item \link{makeGridworld}
\item \link{gridworld}
\item \link{windyGridworld}
\item \link{cliff}
}
}

\section{Bandits}{

\itemize{
\item \link{solveBandit}
}
}

\section{Algorithms}{

Dynamic Programming:
\itemize{
\item \link{evaluatePolicy}
\item \link{iteratePolicy}
\item \link{iterateValue}
}

Model-free Reinforcement Learning
\itemize{
\item \link{td}
\item \link{qSigma}
\item \link{sarsa}
\item \link{expectedSarsa}
\item \link{qlearning}
}
}

