% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reinforcelearn.R
\docType{package}
\name{reinforcelearn}
\alias{reinforcelearn}
\alias{reinforcelearn-package}
\title{Reinforcement Learning in R.}
\description{
Implementations of reinforcement learning algorithms.
}
\section{Algorithms}{


Dynamic Programming:
\itemize{
\item \link{evaluatePolicy}
}

Monte Carlo Methods:
\itemize{
\item \link{predictMC}
}

Temporal-Difference Learning:
\itemize{
\item \link{td}
\item \link{sarsa}
\item \link{qlearning}
}

Function Approximation
\itemize{
\item \link{qlearning2}
}
}

\section{Environments}{

\itemize{
\item \link{gridworld}
\item \link{WindyGridworld}
\item \link{makeEnvironment}
}
}

