% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reinforcelearn.R
\docType{package}
\name{reinforcelearn}
\alias{reinforcelearn}
\alias{reinforcelearn-package}
\title{Reinforcement Learning.}
\description{
Implementations of reinforcement learning algorithms and environments.
}
\section{Environments}{

\itemize{
\item \link{makeEnvironment}
\item \link{Environment}
\item \link{GymEnvironment}
\item \link{MdpEnvironment}
\item \link{Gridworld}
\item \link{WindyGridworld}
\item \link{CliffWalking}
\item \link{MountainCar}
\item \link{MountainCarContinuous}
}
}

\section{Bandits}{

\itemize{
\item \link{bandit}
}
}

\section{Algorithms}{

Dynamic Programming:
\itemize{
\item \link{evaluatePolicy}
\item \link{iteratePolicy}
\item \link{iterateValue}
}

Model-free Reinforcement Learning
\itemize{
\item \link{td}
\item \link{qSigma}
\item \link{sarsa}
\item \link{expectedSarsa}
\item \link{qlearning}
\item \link{actorCritic}
}
}

