% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reinforcelearn.R
\docType{package}
\name{reinforcelearn}
\alias{reinforcelearn}
\alias{reinforcelearn-package}
\title{Reinforcement Learning in R.}
\description{
Implementations of reinforcement learning algorithms.
}
\section{Algorithms}{


Dynamic Programming:
\itemize{
\item \code{\link[=evaluatePolicy]{evaluatePolicy()}}
\item \code{\link[=iteratePolicy]{iteratePolicy()}}
\item \code{\link[=iterateValue]{iterateValue()}}
}

Monte Carlo Methods:
\itemize{
\item \code{\link[=predictMC]{predictMC()}}
\item \code{\link[=MonteCarloES]{MonteCarloES()}}
\item \code{\link[=controlMonteCarlo]{controlMonteCarlo()}}
}

Temporal-Difference Learning:
\itemize{
\item \code{\link[=td]{td()}}
\item \code{\link[=sarsa]{sarsa()}}
\item \code{\link[=expectedSarsa]{expectedSarsa()}}
\item \code{\link[=qlearning]{qlearning()}}
\item \code{\link[=dqlearning]{dqlearning()}}
}
}

\section{Environments}{

\link{gridworld}
}

