% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reinforcelearn.R
\docType{package}
\name{reinforcelearn}
\alias{reinforcelearn}
\alias{reinforcelearn-package}
\title{Reinforcement Learning in R.}
\description{
Implementations of reinforcement learning algorithms.
}
\section{Algorithms}{


Dynamic Programming:

\code{\link{evaluatePolicy}}

\code{\link{iteratePolicy}}

\code{\link{iterateValue}}

Monte Carlo Methods:

\code{\link{predictMC}}

\code{\link{MonteCarloES}}

\code{\link{controlMonteCarlo}}

Temporal-Difference Learning:

\code{\link{TD}}

\code{\link{sarsa}}

\code{\link{qlearning}}
}

\section{Environments}{

\code{\link{gridworld}}
}

