% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reinforcelearn.R
\docType{package}
\name{reinforcelearn}
\alias{reinforcelearn}
\alias{reinforcelearn-package}
\title{Reinforcement Learning in R.}
\description{
Implementations of reinforcement learning algorithms.
}
\section{Algorithms}{


Dynamic Programming:
\itemize{
\item \link{evaluatePolicy}
\item \link{iteratePolicy}
\item \link{iterateValue}
}

Monte Carlo Methods:
\itemize{
\item \link{predictMC}
\item \link{MonteCarloES}
\item \link{controlMonteCarlo}
}

Temporal-Difference Learning:
\itemize{
\item \link{td}
\item \link{sarsa}
\item \link{expectedSarsa}
\item \link{qlearning}
\item \link{dqlearning}
}
}

\section{Environments}{

\link{gridworld}
\link{windyGridworld}
\link{mountainCar}

Link Vignetten!
}

