% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reinforcelearn.R
\docType{package}
\name{reinforcelearn}
\alias{reinforcelearn}
\alias{reinforcelearn-package}
\title{Reinforcement Learning in R.}
\description{
Implementations of reinforcement learning algorithms.
Create an environment using \link{makeEnvironment}.
}
\section{Bandits}{

\link{solveBandit}
}

\section{Algorithms}{

Dynamic Programming:
\itemize{
\item \link{evaluatePolicy}
}

Model-free Reinforcement Learning
Monte Carlo Methods:
\itemize{
\item \link{predictMonteCarlo}
\item \link{MonteCarloControl}
}

Temporal-Difference Learning:
\itemize{
\item \link{td}
\item \link{sarsa}
\item \link{qlearning}
}

Function Approximation
\itemize{
\item \link{qlearning2}
}
}

\section{Example Environments}{

\itemize{
\item \link{gridworld}
\item \link{WindyGridworld}
}
}

