% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reinforcelearn.R
\docType{package}
\name{reinforcelearn}
\alias{reinforcelearn}
\alias{reinforcelearn-package}
\title{A package for reinforcement learning.}
\description{
Implementations of reinforcement learning algorithms.
}
\section{Reinforcements Learning Algorithms}{

\code{\link{evaluatePolicy}}

\code{\link{predictMC}}
}

\section{Environments}{

\code{\link{gridworld_R6}}
}

