% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reinforcelearn.R
\docType{package}
\name{reinforcelearn}
\alias{reinforcelearn}
\alias{reinforcementlearning}
\alias{reinforcelearn-package}
\title{Reinforcement Learning.}
\description{
Implementations of reinforcement learning algorithms and environments.
}
\section{Environments}{

\itemize{
\item \link{makeEnvironment}
\item \link{Environment}
\item \link{GymEnvironment}
\item \link{MdpEnvironment}
\item \link{Gridworld}
\item \link{WindyGridworld}
\item \link{CliffWalking}
\item \link{MountainCar}
\item \link{MountainCarContinuous}
}
}

\section{Policies}{

\itemize{
\item \link{makePolicy}
\item \link{EpsilonGreedyPolicy}
\item \link{GreedyPolicy}
\item \link{SoftmaxPolicy}
\item \link{RandomPolicy}
}
}

\section{Value Function Representations}{

\itemize{
\item \link{makeValueFunction}
\item \link{ValueTable}
\item \link{ValueNetwork}
}
}

\section{Algorithms}{

\itemize{
\item \link{makeAlgorithm}
\item \link{QLearning}
}
}

\section{Extensions}{

\itemize{
\item \link{makeReplayMemory}
\item \link{Eligibility}
}
}

\section{Agent}{

\itemize{
\item \link{makeAgent}
\item \link{getValueFunction}
\item \link{getReplayMemory}
\item \link{getEligibilityTraces}
}
}

\section{Interaction}{

\itemize{
\item \link{interact}
}
}

