% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictMC.R
\name{sampleEpisode}
\alias{sampleEpisode}
\title{Sample episode}
\usage{
sampleEpisode(policy, envir, initial.state)
}
\arguments{
\item{policy}{a policy specified as a probability matrix [states x actions]}

\item{envir}{the environment, a function returning the next state and reward given an action}

\item{initial.state}{integer: the initial state}
}
\value{
a list with sampled actions, states and returns of the episode.
}
\description{
Sample an episode in an environment given a policy. 
Note that this only works for episodic environments 
(e.g. there must be a terminal state).
}
\examples{
set.seed(27)
grid = gridworld_R6$new(shape = c(4, 4), terminal.states = c(1, 16))

initial.state = 3

# make random policy
n.states = nrow(grid$reward.matrix)
n.actions = ncol(grid$reward.matrix)
random.policy = matrix(1 / n.actions, nrow = n.states, ncol = n.actions)

# sample an episode for the random.policy
episode = sampleEpisode(random.policy, grid, initial.state)
print(episode$actions)
print(episode$rewards)
print(episode$states)

}
