% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sarsa.R
\name{sarsa}
\alias{sarsa}
\title{SARSA (Table-lookup)}
\usage{
sarsa(envir, lambda = 0, n.episodes = 100, learning.rate = 0.1,
  epsilon = 0.1, epsilon.decay = 0.5, epsilon.decay.after = 100L,
  initial.value = 0L, discount.factor = 1, seed = NULL)
}
\arguments{
\item{envir}{an R6 class: the reinforcement learning environment
created by \link{makeEnvironment}.}

\item{lambda}{scalar numeric in (0, 1): Then lambda = 0 only
current state is updated (this is equivalent to TD(0)), for
lambda = 1 all states visited are updated, this is roughly
equivalent to every-visit Monte Carlo.}

\item{n.episodes}{scalar integer: number of episodes}

\item{learning.rate}{scalar numeric between 0 and 1: learning rate}

\item{epsilon}{scalar numeric: ratio of random exploration in
epsilon-greedy action selection}

\item{epsilon.decay}{scalar numeric between 0 and 1: decay epsilon
by this factor}

\item{epsilon.decay.after}{scalar integer: number of episodes afer
which to decay epsilon}

\item{initial.value}{scalar numeric: initial values for the action
values Q, set this to the maximal possible reward to encourage
exploration (optimistic initialization)}

\item{discount.factor}{scalar numeric, discounting future rewards}

\item{seed}{scalar integer: random seed}
}
\value{
optimal action value function Q
}
\description{
Sarsa is an on-policy TD control algorithm to find the optimal policy.
The action value function Q will be updated
towards the action value function of the next state S' and next action A'
using an epsilon-greedy policy derived from Q.
The update formula is:
\deqn{Q(S, A) <- Q(S, A) + \alpha[R + \gamma Q(S', A') - Q(S, A)]}
}
\examples{
grid = WindyGridworld$new()
WindyGridworld1 = makeEnvironment(transition.array = grid$transition.array, 
  reward.matrix = grid$reward.matrix, 
  initial.state = 30L)
res = sarsa(WindyGridworld1, n.episodes = 100, seed = 123)

}
\references{
Sutton and Barto (Book draft 2016): Reinforcement Learning: An Introduction
}
\seealso{
\link{td}

\link{qlearning}
}
