% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeGridworld.R
\name{windyGridworld}
\alias{windyGridworld}
\title{Windy Gridworld}
\usage{
windyGridworld()
}
\value{
[\code{R6 Class}] \cr
Reinforcement Learning environment.
}
\description{
Windy Gridworld problem for reinforcement learning. Actions include
going left, right, up and down. In each column the wind pushes you up a
specific number of steps (for the next action). If an action would
take you off the grid, you remain in the previous state. For each step you
get a reward of -1, until you reach into a terminal state.
}
\details{
This is the gridworld (goal state denoted G, start state denoted S).
The last row specifies the upward wind in each column.
\tabular{rrrrrrrrrr}{
 . \tab . \tab . \tab . \tab . \tab . \tab . \tab . \tab . \tab . \cr
 . \tab . \tab . \tab . \tab . \tab . \tab . \tab . \tab . \tab . \cr
 . \tab . \tab . \tab . \tab . \tab . \tab . \tab . \tab . \tab . \cr
 S \tab . \tab . \tab . \tab . \tab . \tab . \tab G \tab . \tab . \cr
 . \tab . \tab . \tab . \tab . \tab . \tab . \tab . \tab . \tab . \cr
 . \tab . \tab . \tab . \tab . \tab . \tab . \tab . \tab . \tab . \cr
 . \tab . \tab . \tab . \tab . \tab . \tab . \tab . \tab . \tab . \cr
 . \tab . \tab . \tab . \tab . \tab . \tab . \tab . \tab . \tab . \cr
 . \tab . \tab . \tab . \tab . \tab . \tab . \tab . \tab . \tab . \cr
 . \tab . \tab . \tab . \tab . \tab . \tab . \tab . \tab . \tab . \cr
 0 \tab 0 \tab 0 \tab 1 \tab 1 \tab 1 \tab 2 \tab 2 \tab 1 \tab 0 \cr
}
}
\examples{
env = windyGridworld()

}
\references{
Sutton and Barto (Book draft 2017): Reinforcement Learning: An Introduction Example 6.5
}
